New version based on a more recent master of the previous code base.
The initial (parent) commit corresponds to the code submitted as part
of "[CQ 7368] New: Sirius Initial Contribution" (see
http://dev.eclipse.org/ipzilla/show_bug.cgi?id=7368). It was based on
a snapshot of the previous code base (Viewpoint) as of 2013-06-23.

Since then the Viewpoint code base has moved on with bug fixes and new
features, so this second commit updates Sirius based on a more recent
version of Viewpoint (from 2013-08-23).

Several branches of the Viewpoint code base are still active, so in
the next few weeks/months we will re-integrate into Sirius all the
fixes and improvements done there, until all active development has
fully transitioned to Sirius.

In addition to many bug fixes and a few new features, the most visible
change between the two versions is the addition of the documentation
plug-in (org.eclipse.sirius.doc), which was not included in the
initial submission.

Signed-off-by: Pierre-Charles David <pierre-charles.david@obeo.fr>