[build] Fix Tycho build on all supported platforms.
This required somes tweaks in the target platform definitions:
* be more precise about the versions of Xtext to includefor each
  plarform version;
* use version 4.2 of the platform for Juno instead of 3.8.

Also the hard-coded requirement on platform and EMF features in the
org.eclipse.sirius.runtime feature definition was removed (it made the
Juno build, and only it, fail for reasons I have not understood yet).

To launch a build, simply issue the following command in the root
folder of a clone:

 mvn -f packaging/org.eclipse.sirius.parent/pom.xml \
   -Dplatform-version-name=juno clean package

where platform-version-name can be one of: helios, indigo, juno,
kepler.

Signed-off-by: Pierre-Charles David <pierre-charles.david@obeo.fr>