[build] Simplify the parent POM by removing profiles definitions.
After the extraction of target platforms into external files which can
be referenced using ${platform-version-name}, the only remaining
difference between the different Maven profiles in the parent POM was
the Xtext support modules, which were present in in all profiles
except Helios.

It turns out that adding the proper version of Xtext in the Helios
target platform and relaxing the version requirement in
org.eclipse.sirius.common.xtext to Xtext 2.1.1 instead of 2.2.1 seems
to work (i.e. it builds successfully, I have not tested the result).

Adding the Xtext modules to the Helios profile means all the profiles
definitions are exactly the same, and thus they can be removed (and
the Xtext modules moved up into the main modules list).

Signed-off-by: Pierre-Charles David <pierre-charles.david@obeo.fr>