[build] Simplify the update-sites structure.
For the same reasons we used to have one profile per target platform
(Xtext not included in the helios build), we also had one update-site
definition per target. And for the same reasons we do no need the
profiles anymore we also can use a single update-site definition.

Replace the 5 update-sites definitions (4 of which were identical)
with a single one (one of the 4).

Signed-off-by: Pierre-Charles David <pierre-charles.david@obeo.fr>