[417812] enable signing and pack2000 using CBI
When the profile "sign" is enabled then the CBI
eclipse-jarsigner-plugin will communicate with the eclipse
servers to sign every JAR.

This is set in the parent pom so that it has effects and every jar
produced by the build.

Please note that the sign profile is disabled by default as we only
want it "on" when building on the Eclipse servers.
It can be enabled either by explicitely enabling the profile "sign" or
by setting the BUILD_SIGN environment variable to "true".

Note also that the signing and pack2000 will only be done
during the *verify* phase of the maven build.

Change-Id: I9fedf352cfc2b684a1a1586eb18e927be0eab114