Post-rebranding fixes.
The automatic rebranding operation performed on the previous code base
for the initial Sirius submission produced working code but with the
wrong names and/or locations for many elements, especially for the
types and packages associated with the metamodel.

This fixes a lot of these issues (but many remains). In particular:
- it reverts the name of the "Viewpoint" concept (and associated
  references) in the core metamodel, as it was wrongly renamed into
  "Sirius" before;
- it changes the generation directory/package for the core metamodel
  into org.eclipse.sirius.viewpoint (instead of just
  org.eclipse.sirius);
- it ensures a clean (idempotent modulo a "Source > Format" and
  "Source > Organize Imports" step) regeneration of all the code
  generated from all our metamodels (viewpoint.ecore,
  contributions.ecore, sequence.ecore, table.ecore, tree.ecore).

The post-rebranding work is not finished, by far, this is seems stable
enough to serve as a new reference state for further work.

Signed-off-by: Pierre-Charles David <pierre-charles.david@obeo.fr>