[422478] Migrate VSM and representations file from Viewpoint to Sirius
This commit adds new migration participants:
* One which maps nsURIs of all the Sirius EPackage with their current
value.
* Another which replaces type of Diagram (notation.Diagram#type: Sirius
replaces Viewpoint)

In doing so, the IMigrationParticipant contract is revised so that
a participant can provide an EPackage instance from a nsURI.

The version in the file (odesign or aird) does not corresponds to a
Sirius version but more to a file format version (7.0 file format
corresponds to Sirius 0.9)
The nsURI target strings have been hardcoded because this migration must works even with Sirius 2.0 in which the nsURI field on the EPackage implementation may have changed by then.

Change-Id: I00b48f346688eafc9773a8f7484d645ece2acaea