Use an external reference for the JARs needed to build the doc.
We used to include the required JARs directly inside the "lib"
directory, but the Eclipse rules prevent this so that directory was
removed before submission.

Use the same technique for the EEF bridge as for the same doc: we
assume the use has installed the JARs locally in his HOME folder.

Change-Id: I03f7c4cfe8038d7d9b4f707a0647c404ffd84571
Signed-off-by: Pierre-Charles David <pierre-charles.david@obeo.fr>