Remove the 'specific' source folder from org.eclipse.sirius.diagram
The folder was a left-over from the time when the 'src' folder
contained code generated by GMF Tooling and 'specific' contained
manually created code. We abandonned the possibility to regenerate the
GMF part using GMF Tooling a long time ago, so the separation does not
make sense anymore.

Note that classes which used to be generated by GMF Tooling can still
be identified if required using the '@was-generated' JavaDoc
annotation.

Change-Id: I2284733854492283bf8b51522716796540711095
Signed-off-by: Pierre-Charles David <pierre-charles.david@obeo.fr>