[421616] Remove the basic combo for the LabelFormat property
The BasicLabelStyleDescription.labelFormat property is implemented
using an enumerated type (FontFormat), which by default are shown
using combo boxes in the property sheets. In this particular case a
set of two check boxes better match the semantics of the property, so
we have a PropertySection which implements the check boxes UI (see
BasicLabelStyleDescriptionLabelFormatPropertySection), so we should
hide the combo-based property section.

Change-Id: Ie87a27467883ecf0cee69cfd4b3eecb17b8f807d
Signed-off-by: Pierre-Charles David <pierre-charles.david@obeo.fr>