[cleanup] Update all @since annotations to reference 0.9.0
The versions which were referenced do not make sense in the context of
Sirius; they corresponded to versions of individual Viewpoint bundles
(which did not even match the global Viewpoint version numbers).

In the context of Sirius, all the APIs are available since the first
version, which is 0.9.0.

Note that this commit does not add "@since 0.9.0" to all APIs, but
only updates existing annotations.

Change-Id: I5a95e3094c5aad5e20e9abe482e8e59e324a27db
Signed-off-by: Pierre-Charles David <pierre-charles.david@obeo.fr>