[423037] Regen without any changes.
This commit removes all the contents of src-gen folder and generate
without any changes in order to have a clean base of code (including the
javadoc contents and formatting). This step will allow us to facilitate
the comparison after the splitting.
All the generated files have been compared with the deleted ones to
restore the "@not-generated" code.

Change-Id: I1e8dd2aeca1e61ac8b2a8d5b568822faebc63f6d
Signed-off-by: Laurent Redor <laurent.redor@obeo.fr>