[423037] Regen org.eclipse.sirius.ui without any changes.
This commit removes all the contents of src-gen folder and generate
without any changes in order to have a clean base of code (including the
javadoc contents and formatting). This step will allow us to facilitate
the comparison after the splitting.
All the generated files have been compared with the deleted ones to
restore the "@not-generated" code.
This commit also cleans the "icons/full/obj16" folder: The unused icons
have been deleted (and the associated entries in licence.txt file).

Conflicts:
	plugins/org.eclipse.sirius/model/viewpoint.genmodel

Change-Id: I669e71dc7d56fa415d13c88bd529d293d6149534
Signed-off-by: Laurent Redor <laurent.redor@obeo.fr>
Signed-off-by: Laurent Redor <laurent.redor@obeo.fr>