[423037] Regen org.eclipse.sirius.editor without any change.
This commit removes all the contents of src-gen folder and generate
without any change in order to have a clean base of code (including the
javadoc contents and formatting). This step will allow us to facilitate
the comparison after the splitting.
All the generated files have been compared with the deleted ones to
restore the "user" code.

Change-Id: I37e05b5ffed3d29b6731edfc890e8bbe4dd96dcc
Signed-off-by: Laurent Redor <laurent.redor@obeo.fr>