[423037] Regen org.eclipse.sirius.editor.sequence without any change.
This commit removes all the contents of src-gen folder and generate
without any change in order to have a clean base of code (including the
javadoc contents and formatting). This step will allow us to facilitate
the comparison after the splitting.
All the generated files have been compared with the deleted ones to
restore the "user" code.

Change-Id: I2dc0e50582fbf3ab1d182bfb4c8cab5fba02a9f3
Signed-off-by: Laurent Redor <laurent.redor@obeo.fr>