[425561] Wire a DanglingRefRemoval trigger
With this commit removing dangling reference does not need
to be explicit anymore and always use the special logic needed
for some references.

The dangling reference to ignore predicates have been moved to the
DanglingRefRemovalTrigger and reused in the other classes.

An additional method has been added in RemoveDanglingReferences to
manage the resource case.

Change-Id: I624fe0298cf2fc245fd2c6c3828784e5954b0c19