[426687] Optimize session's modification tracking
This replaces the current, costly setTrackingModification(true) with a
more efficient ModelChangeTrigger.

Sirius is forcing the usage of setTrackingModification on the resources.
It is using the isModified information to know if the session is in
sync. We are using this information to know whether the session is dirty
or not, and we know some adopters are using this information too. Not
having this boolean value anymore is not something we can target now.
But, we can avoid installing "yet another adapter" by replacing this
with a ModelChangeTrigger (which will be called by our "mighty
adapter").

This speeds up the Session opening and avoid filling the eAdapters()
list with on more adapter, and we often loop over this one for finding a
specific type instance.

Furthermore this make this choice, of having the modification tracking
enabled, a bit more configurable.

At last but not least, if our "mightly adapter" gets lazy, that's one
less step to do.

Please note that GMFResource forces the setTrackingModification itself
leading to .aird resources always having one anyway.

This commit introduces an API break in DAnalysisSessionImpl with a
protected method which disappeared.

Bug: 426687
Change-Id: I78098ac6b2e4b41880326d7242594bfb08b9546d
Signed-off-by: Pierre-Charles David <pierre-charles.david@obeo.fr>