[425561] Add consitency in delete behaviors after dangling improvements
Correcly do the delete table/tree element without delete tool. It was
done by the "auto refresh" even in manual refresh mode. A next step will
be to correcly handle the manual refresh in
AbstractCommandFactory.addRefreshTask to be consistent with the
AbstractCommandBuilder. AbstractCommandFactory handle the tool==null
case as a force refresh whereas the command builder do not set the force
refresh when there is no tool.

Change-Id: I13a583d32758ff53c8a88de38fdcee8291c56fe9
Signed-off-by: Maxime PORHEL <maxime.porhel@obeo.fr>