[425561] Change the RefreshEditorsPrecommitListener into a ModelChangeTrigger
This allows to correctly manage the order/sequence of precommit actions:
remove dangling refs, refresh opened editors, refresh the GMF models,
compute the filter applications, layout the sequence diagram,...

The old RemoveDanglingReferenceTasks were done during the command, then
the RefreshEditorPrecommitListner (which was created before the
SessionEventBroker) did a refresh when needed and then the
SessionEventBroker was launching its ModelChangeTrigger precommit
commands. 

We could init the RemoveDanglingRefTrigger after the
addRefreshEditorsListener() call, during the session opening , but then
the remove dangling ref will be done after the refresh (it could also
introduce several refresh, removedangling, sequence layout, refresh, ..
cycles). If we init the trigger before, all model trigger actions would
be done before the refresh, then the refresh will be done and then the
model change trigger will done a second time. This have a major impact
on Sequence where the layout will be done before and after the refresh
with unwanted effects.

The clean solution is to transform the RefreshEditorsPrecommitListener
into a model change trigger with a priority between the
DanglingRefRemovalTrigger and the GMF canonical trigger.


Change-Id: Iad0c727c987f3cf2d6b845091474e88d69392b1e
Signed-off-by: Maxime PORHEL <maxime.porhel@obeo.fr>