[427803] Use VisibilityPropagatorAdapter for opened editor and repair
We do not need to create/install this adapter on the DAnalysis for all
diagrams. We should do it only when needed: when a diagram editor is
opened or during repair. 

The GMFDiagramUpdater (the name should be changed) is both created on
diagram opening and for repair. So it is the perfect handler for the
VisibilityPropagatorAdapter.

Furthermore, this commit will facilitate the split between Diagram and
Core by releasing one dependency between the DAnalysisSessionImp and a
diagram only related concept.

Change-Id: Ibf96242f1edaca6b62626823da9ea73552bc0622
Signed-off-by: Maxime PORHEL <maxime.porhel@obeo.fr>