[428756] Improve tree refresh algorithm
We have performance issue on tree editor in manual refresh.
We have performance issue on 2 cases :

1. On expand, because in DTreeViewerListener we execute a
TreeFoldingRecordingCommand which refreshes all the subtrees of the
expanded DTreeItem even for DTreeItem contained in a collapsed
DTreeItem.
In addition if we are in auto refresh or in manual, when the tree is
expanded or collapsed, we call always RefreshTreeElementTask which
refreshes the expanded
DTreeItem, this is not done by the TreeFoldingRecordingCommand which
refreshes only the subtree. RefreshTreeElementTask also refreshes the
style
of all expanded sub trees items in case of expanded DTreeItem, and
refreshes
all the DTree in case of expand of DTree.

Note that RefreshTreeElementTask which is executed in a SiriusCommand
executes RefreshRepresentationsCommand/RefreshTreeElementCommand in a
CommandStack which makes no sense.

2. On item creation, as for TreeItem expansion, DTreeRefresh refresh all
the subtree
even for DTreeItem contained in a collapsed DTreeItem.

3 changes in the patch:
1. org.eclipse.sirius.synchronizer.CreatedOutput.synchronizeChildren()
method added to know if we must refresh the subtree of a
DTreeItemContainer, this allows to do lazy refresh for collapsed
DTreeItem. But to have DTreeContentProvider.hasChildren() returning true
for container having children not yet created, we create only one
children, without refreshing it as it will be refreshed when user will
expand the DTreeContainer containing it.
2. TreeItemContainerChildSupport.reorderChilds() changed to not call
ECollections.sort() when there is only one mapping in place.
3. DTreeViewerListener.treeExpanded() changed to not use
RefreshTreeElementTask when we are in auto-refresh.

To leave the opportunity to do a full refresh without lazy we add a
boolean to the Sirius refresh api to allow to do it.

Bug: 428756
Change-Id: I1b9a03d5250060611c2ddc012d1310999a3f4c8e
Signed-off-by: Esteban Dugueperoux <esteban.dugueperoux@obeo.fr>