[428756] Improve tree refresh algorithm
We have performance issue on tree editor in manual refresh.
We have performance issue on 2 cases :

1. On expand, because in DTreeViewerListener we execute a
TreeFoldingRecordingCommand which refresh all the subtree of the
expanded DTreeItem even for DTreeItem contained in a collapsed
DTreeItem. In addition if we are in auto refresh or in manual, we expand
or collapse, we call RefreshTreeElementTask which refresh the expanded
DTreeItem, which is not done by the TreeFoldingRecordingCommand which
refresh only the subtree. RefreshTreeElementTask also refresh the style
of all expanded sub tree item in case of expanded DTreeItem, and refresh
all the DTree in case of expand of DTree.

Note that RefreshTreeElementTask which is executed in a SiriusCommand
execute RefreshRepresentationsCommand/RefreshTreeElementCommand in a
CommandStack which makes no sense.

2. On item creation, as for expand DTreeRefresh refresh all the subtree
even for DTreeItem contained in a collapsed DTreeItem.
I will submit a first release of patch to gerrit to improve refresh
performance.
3 changes in the patch:
1. org.eclipse.sirius.synchronizer.CreatedOutput.synchronizeChildren()
method added to know if we must refresh the subtree of a
DTreeItemContainer, this allow to do lazy refresh for collapsed
DTreeItem. But like it is now done lazily, we must also update the
DTreeContentProvider.hasChildren() to not call getChildren() be have the
info that there is children otherwise. For that we use the
HasChildrenNotYetSynchronizedMarker to indicate that a
DTreeItemContainer has children.
2. TreeItemContainerChildSupport.reorderChilds() changed to not call
ECollections.sort() when there is only one mapping in place.
3. DTreeViewerListener.treeExpanded() changed to not use
RefreshTreeElementTask when we are in auto-refresh.

Change-Id: I8ba4e4cf85e0abcc74ace8f418fc1c398feec83b