[428976] Use Guava's Cache API properly
The previous change from the MapMaker.makeComputingMap() (which does not
exist in recent versions of Guava) to the Cache API was broken. Simply
calling asMap() on a Guava cache gives a view on the (empty) cache, but
does not trigger the addition of new entries on get(). The new version
explictly uses the LoadingCache API and deals with the possible
exceptions.

Change-Id: Ie8dd3814d191cacf72c200534c081d81d0956f58
Signed-off-by: Pierre-Charles David <pierre-charles.david@obeo.fr>