[426806] Remove not implemented setAutoRefreshXXX methods
The methods were not actually implemented, so it can safely be removed
along with the code which tried to keep the setting synchronized with
the corresponding preference.

The only concrete class implementing the interfaces were
TableCommandFactory and TreeCommandFactory, but they were not real implementation: the setAutoRefreshXXX() method just did:

// TODO Feature not implemented.

The corresponding method has already been removed in
IDiagramCommandFactory.

Change-Id: I357304bd46778882ae385f9f5bcb8d40db9673c0
Signed-off-by: Maxime PORHEL <maxime.porhel@obeo.fr>