[build] Add a staging target platform definition
Our current sirius_luna.target* point to releases/luna, and thus (at
this time) only see Luna M5. To ensure we are compatible with the
upcoming Luna M6 we need to build against releases/staging, where
M6+0/+1/+2 are being published currently (we are at +3).

This has allowed us to see we were not compatible with the version of
Guava (v15) which we will get in the final M6 (this is fixed, see
https://git.eclipse.org/r/#/c/23233/).

I also widened the version ranges in sirius_luna so that it will
continue to resolve correctly after M6 is released.

Change-Id: I9c199a72f6c17308e5c3f384639578f975736577
Signed-off-by: Pierre-Charles David <pierre-charles.david@obeo.fr>