[cleanup] Replace broken implementation of findDiagramElements()
As mentioned in the code comment, the method is broken (it depends on
the presence of an interpreter for a legacy language not present by
default in Sirius), not used anywhere in the current code base, and
deprecated anyway. The new version has the same behavior but at least is
fast and does not pretend to do something.

The method will be removed when the corresponding metamodel operation is
deleted as part of bug #427247. In the meantime this cleanup at least
removes hard-coded references to a legacy language not supported by
default.

Change-Id: I2702fd1a4e21ad0358d126d0584b06931ee46f65
Signed-off-by: Pierre-Charles David <pierre-charles.david@obeo.fr>