[427803] Fix Diagram Plugin issues
DiagramPlugin is is now a standard EMFPlugin (as SiriusPlugin,
TablePlugin, ...). 
It was previously an AbstractUIPLugin, DiagramUIPlugin is now used as
the preference store container for all diagram preferences. These
preferences still mix UI and non-UI aspects in classes from both diagram
plugins, a proper separation will be done in Bug 430676.

Bug: 427803
Change-Id: I26a7feb8e19645d00bf5f89322751be14de3d805
Signed-off-by: Maxime PORHEL <maxime.porhel@obeo.fr>