[431390] Improve Refresh on opening preference management
The old DialectManager.setRefreshOnOpening() could have let a developer
think that he had the possibility to programmatically change this
behavior, but this was called directly on reaction to changes in the
preference page without any checks. The dialect manager was updated from
a preference change listener added during SiriusEditPlugin.start(). 

Furthermore this corresponds to UI behavior and preference. 
The isRefreshActivatedOnRepresentationOpening() method has been moved to
DialectUIManager which directly call the preference service. 

Bug: 431390
Change-Id: I0a4f23b244801c816d18323f1f957f64b7c1efb2
Signed-off-by: Maxime PORHEL <maxime.porhel@obeo.fr>