[431634] Improve tabbar contributions
Merge SetStyleToWorkspaceImageContributionItem and
DeleteFromDiagramContributionItem in SiriusActionBarContributionItem
(used in old ui, to populate the main action bars). 
Use the TabbarActionContributionItem to add the correspondign actions in
the tabbar: it uses the part selection service and not the page's one.
Modify the TababrActionContributionItem to add the capability the action
enablement through selection change propagation.

Bug: 431634
Change-Id: I8677f209a6a40fb7cec42f9b3e129c288e41dc6e
Signed-off-by: Maxime PORHEL <maxime.porhel@obeo.fr>