[427017] Remove AirdCrossReferenceAdapter
This was a "content adapter" which installs itself on all elements of
the resource. It was made to customize GMF's own
org.eclipse.gmf.runtime.emf.core.util.CrossReferenceAdapter.

However considering the way we use GMF, we do not need this adapter at
all. To detect when GMF install it (as a collateral damage of the use of
a command, policy, API, ...), some checks have been added in the
abstract JUnit and SWTBot test cases. Thanks to this, our calls to
ViewUtil.destroy(View) have been detected, these calls are not needed
with Sirius: a simple EcoreUtil.remove() is sufficient in the detected
case, then the DanglingRefRemovalTrigger will do the same cleanup than
ViewUtil but without the installation of the GMF CrossReferencerAdapter.

Bug: 427017
Change-Id: I0fe6e43b8ecff1334e29e50ca550dfea37a1e520
Signed-off-by: Maxime PORHEL <maxime.porhel@obeo.fr>