[427247] Remove DNode.ownedDetails and DDiagramElementContainer.ownedDetails
Both references where defined in the metamodel but not actually used
anywhere. It seems they were added to support "sub-diagrams", but this
was either never implemented or removed a long time ago. Navigation
links support the same intended use cases with no restriction on the
structure of the diagrams and none of the performance impacts of putting
a whole diagram inside another.

Bug: 427247
Change-Id: I80b764a90660cb16cc5514a7eb00be543896c6c5
Signed-off-by: Pierre-Charles David <pierre-charles.david@obeo.fr>