[test] Improve SWTBotSiriusDiagramEditor.zoom()
With a luna target,
designerBot.toolbarSpecialDropDownButtonWithTooltip("&Zoom"); seems to
randomly throw a WidgetNotFoundException, after a 10 second timeout. It
slows down the tests. 

There is no real need to do this with the zoom combo from the
tabbar/action bar, we can directly change the zoom value of the zoom
manager. 

A new method zoomFromToolbar have been added to for tests which want to
explicitely change the zoom from the toolbar.

Change-Id: I29dfebc8bd41f244965423f71bfc9311c2800133
Signed-off-by: Maxime PORHEL <maxime.porhel@obeo.fr>