[427803] Correct Split merge side effects
The AirNoteAttachmentEditPolicyProvider declaration in diagram.ui
plugin.xml was reintroduced in the merge commit too. It has been
intially removed in commit
8db397b66ee5479a5613d5987663589e14128510 but during the diagram split
it was moved to the diagram.ui plugin and the merge commit
f47d4f72102a795cedc92334593480d56d72b44d was not able to delete the
correcponding extension.

Change-Id: Ibb6f11cf34c550324465f155b057654945c5420d
Signed-off-by: Maxime PORHEL <maxime.porhel@obeo.fr>