[430263] Replace the Sirius decorator by a more complete LabelProvider.
The current Sirius decorator ("org.eclipse.ui.decorators" extension
point) implies that if Sirius is installed, a large part of Sirius
plugins are getting loaded and started even if your are not using it.
This decorator adds a decorator on semantic resources of non Modeling
project. This is now done by the SiriusCommonLabelProvider.

Warning: If the Sirius Content Management is active (for example in
"Project Explorer"), Sirius plugins are getting loaded and started even
with this fix.

Bug: 430263
Change-Id: I7cc11ec343b54a12a58531c5eb366a88d5843987
Signed-off-by: Laurent Redor <laurent.redor@obeo.fr>