[430263] Remove the ModelingProjectManagerStartup startup extension
This extension implies that if Sirius is installed, a large part of
Sirius plugins are getting loaded and started even if your are not using
it.

The sessionManagerListener of ModelingProjectManagerImpl, previously
added by this extension point, is now added during the loading of the
first session.
The initialization of a modeling project (detect if the modeling project
is valid, ie with one and only one representations file) is now done at
the first access of this project and not at the Eclipse start.

Warning: If the Sirius Content Management is active (for example in
"Project Explorer"), Sirius plugins are getting loaded and started even
with this fix.

Change-Id: I3c3c3891bca0f2531935931b8ad97abaef01b320
Signed-off-by: Laurent Redor <laurent.redor@obeo.fr>