[430263] Remove the ModelingProjectManagerStartup - Clean step
This startup extension implies that if Sirius is installed, a large part
of
Sirius plugins are getting loaded and started even if your are not using
it.

This commit is a first cleaning step to dissociate the core of this
change:
* Change behavior when a file is removed in a modeling project: Now
there is a validation launched when a reresentations file is deleted in
an invalid project.
* The method ModelingProject.getMainRepresentationsFileURI now marks
this project as invalid and adds a marker on it, if it is considered as
invalid (no or many representations files).
* The method InitializeModelingProjectJob.initializeModelingProjects()
does not add marker anymore (it is done by
ModelingProject.getMainRepresentationsFileURI()).

Change-Id: I9ac762e5a99867dfc592f0b43b51277507b20d86
Signed-off-by: Laurent Redor <laurent.redor@obeo.fr>