[429682] Introduce EReferencePredicate to replace Predicate<EReference>
Introduce a new interface EReferencePredicate, functionally equivalent
to Guava's Predicate<EReference>, and replace all instances of the Guava
type in our APIs with the new type.

The new interface is voluntarily ad-hoc and not a generic Predicate<T>
to avoid going in the direction of reproducing parts of Guava in our own
code base.

Bug: 429682
Change-Id: Iaea4f112f58c23f4a755542528c3072a2a40cac3
Signed-off-by: Pierre-Charles David <pierre-charles.david@obeo.fr>