[cleanup] Remove the 60s timeout in SWTBot session opening
The previous code was not clear, the use of CountDownLatch.await was
done after the waitProgressMonitorClose("Open session") which has been
added during the introduction of a progress monitor in the session
opening.
Now, after the closing of the progress monitor dialog "Loading models",
we simply do two "wait until" (with the default swtbot timeout: 10s), to
wait that the selected session is added to the session manager and is
open.
The condition will throw a timeout if no new session has been opened.
The UIPerspective will not try to take the last opened session any more.

Change-Id: Id3621119c06b5f007571615beaa726b234bda3e9
Signed-off-by: Maxime PORHEL <maxime.porhel@obeo.fr>