[429682] Remove usage of Predicate in IDiagramDescriptionProvider
Several methods in IDiagramDescriptionProvider used to return a
Predicate<DDiagramElement> that the client code then had to invoke.
Change this to directly take the DDiagramElement to test and return the
resulting boolean.

Also change the related PinHelper in the same way.

The resulting code is a little awkward in places as I tried to reduce
the amount of changes in order to facilitate the review (i.e. sometimes
predicates are still created internally for no good reason). If this is
accepted, some cleanup should then be done.

Bug: 429682
Change-Id: I8d951c0ae8f837e63776f542fa4dcbb2c3503e21
Signed-off-by: Pierre-Charles David <pierre-charles.david@obeo.fr>