[432931] Force option to avoid Xtext concrete syntax validation
Add a specific saving policy (XtextSavingPolicy) which is wired in the
sessions using a Session manager listener to force the use of the Xtext
option to ignore concrete syntax validation during save.

This change expose a new method: Session.getSavingPolicy() which is used
by the XtextSavingPolicy to get the previous one in order to be able to
delegate to it.

Bug: 432931
Change-Id: Idc43ab8648d40abfbf23ef3354e1665ae863650b