[430579] Use a better default candidates expression for tree items
The default value for TreeItemMapping's semanticCandidatesExpression was
the empty string, which means "any compatible element in the
ResourceSet" and can be very costly on large models. Use the same value
as for node mappings by default, feature:eAllContents (see
ContainerMappingItemProvider.collectNewChildDescriptors()).

The change only applies when using the VSM editor, and does not change
the default value in the metamodel. It does not impact existing VSMs and
thus requires no migration. Users who want the old behavior can simply
remove the automatically filled expression in the properties view.

Bug: 430579
Change-Id: Iea7a48f9c23d0799fcae1dfaa7dac2c84c90e7df
Signed-off-by: Pierre-Charles David <pierre-charles.david@obeo.fr>