[430536] Add a new TP with bleeding-edge versions of all our dependencies
The new sirius_canary.target (and the corresponding .targetplatform)
reference the latest nightly or last successful build of all the Eclipse
projects we depend on. The goal is to use this as an early warning
system of breaking changes, to avoid discovering them in the rush of a
release (the canary in the coal mine).

Note that most of the dependencies have a version="0.0.0" (version=lazy
in the .targetplatform file) constraint so that we really always get the
very latest available. There is one exception and one issue with this:
* the exception is the platform itself. It seems that the way its
  update-site is structured, setting a version="0.0.0" causes all kinds
  of problems when resolving the target. This means the .target file is
  always bound to a specific platform build, and we must regularly
  refresh it from the .targetplatform definition to be sure we use the
  latest I-build.
* the other issue is that it means we always get the version with the
  maximum version number, which is generally fine but does not always
  correspond to the latest in the rare cases where a project reverts its
  current version number (like GEF did recently, going back to version
  3.9.x for master after publishing builds versioned as 3.10 from that
  branch).

Bug: 430536
Change-Id: Ia24f7ee68680e83dce140740f4054f00a5bdd7df
Signed-off-by: Pierre-Charles David <pierre-charles.david@obeo.fr>