[432986] Enable EMF's child extender mechanism
When moving the diagram metamodel out of the core, we lost some creation
items in the VSM editor's context menus, as they moved into an EPackage
the core genmodel does not know about. Sirius already provides several
ad hoc mechanisms to handle this kind of things (see
DialectUISercice.provide*() methods), but none of them correspond to the
locations that need to be extended in this case.

Instead or creating yet another ad hoc extension method (or two), enable
EMF's native mechanism to handle this. It should probably be enabled
everywhere and replace our custom mechanisms, but for now just do the
minimum:
* enable "Extensible Provider Factory" in all the description EPackages
  in viewpoint.genmodel so that the item providers look for extensions;
* enable "Child Creation Extenders" in all the description EPackages in
  diagram.genmodel so that the extensions for viewpoint.ecore are
  properly generated;
* remove the custom contributions which were added in
  DiagramDialectUIServices.provideNewChildDescriptors(), as they are now
  redundant and create duplicate entries in the menus.

The patch is rather large, but the only things manually changed are a
few boolean in the 2 genmodels and the change in
DiagramDialectUIServices.provideNewChildDescriptors(). All the rest is
the result of the EMF re-generation.

With this change, we get back the 4 menu items which were identified as
missing compared to M6 (before the split). See the bugzilla for the
reasons why we think they are the only ones.

Bug: 432986
Change-Id: I5735dc63f75c7292af1ee7640fc2a39399ea293d
Signed-off-by: Pierre-Charles David <pierre-charles.david@obeo.fr>