Fix NPE when opening a diagram
There is quite often the case in EcoreTools that some resource targets
another resource which is not in the ResourceSet. It got resolved
directly using the EPackage registry. In those cases I often ends up
being unable to open a diagram because of an NPE: Sirius use the current
EObject to retrieve the editing domain and the current EObject in this
case might be an EObject from Ecore.ecore itself.

In particular
SubDiagramDecorator.shouldHaveSubDiagDecoration(DRepresentationElement)

will query for associated representations on every instance which is
getting displayed, in some case EAttribute for instance.

With this change the call will not fail with an NPE and will retrieve
the representations.

Change-Id: Ie8dd90c9a84dbca28f019360fbffd279eddb3f35