[431359] Move the definition and implementation of aird resources into the core plug-in
For historical reasons the session resources (*.aird) depend on GMF and
are defined in the plug-ins for the diagram dialect. However some of the
code in the core (not diagram-specific) plug-in needs to load these
models (at least to correctly identify them as session resources). This
fails if the core plug-ins are installed but not the diagram-specific
ones.

This can happen rather easily since for example a Viewpoint
Specification Project created with our own wizard depends on
org.eclipse.sirius but not the diagram part. If a user installs such a
plug-in, p2 will only get the required core plug-ins, not the whole
feature with the diagram-specific parts.

Move all the required classes into the org.eclipse.sirius core plug-in.
It makes that plug-in depend on parts of GMF runtime, at least until we
can properly remove all the dependencies towards GMF, but that is less
severe than exposing users to potential crashes.

Bug: 431359
Change-Id: I4bdaea9c2e387c13549b845abaf0ba93a2d7de64
Signed-off-by: Pierre-Charles David <pierre-charles.david@obeo.fr>