[439103] Completion on Acceleo 3 expressions is not available anymore in odesign file
A NPE is thrown on interpreted expression in odesign completion when a
Java extension to an empty Java class is defined. The
CompilationResult.getStatus() returns null. The fix consists to consider
CompilationResult.getStatus()==null as Status.OK in the condition where
there is the NPE.

This is a backport of the fix for bug #433821 for the v1.0.x branch.

Bug: 439103
Change-Id: Ie0a594cb3bb6196932cf9ef57fed8be5d010d17f
Signed-off-by: Esteban Dugueperoux <esteban.dugueperoux@obeo.fr>
Signed-off-by: Pierre-Charles David <pierre-charles.david@obeo.fr>