[434099] remove usages of SiriusUtil.isFromSirius
SiriusUtil.isFromSirius is checking for the class namespace
to return true. This has strong implication regarding extensibility
as any code using this predicate will filter subclasses or sub
interfaces of Sirius which are not in the org.eclipse.sirius namespace.

This commit replace the internal usages of isFromSirius() by instanceof checks
which will return true for all the subtypes.
It introduces a "marker" Interface ISiriusEditPart to do so.

Change-Id: I82878400f78234b03ce8733c98439f830ee1aff3