[437646] Enhance AirDCrossReferenceAdapter installaton
Have AirDCrossReferenceAdapterImpl installed on ResourceSet and be
adapted recursively only on aird resource, i.e.
ResourceQuery.isRepresentationsResource().
Also remove it on session close as it is done for the semantic cross
reference adapter. Move this specific code in other methods to avoid
checkstyle error in DAnalysisSessionImpl.close() about number of lines.

Bug: 437646
Signed-off-by: Esteban Dugueperoux <esteban.dugueperoux@obeo.fr>
Change-Id: I16ec7b4e47d111c97f87c5df3d3ea55afa0a0aa5