[437097] Update policies concerned by this feature
* Use of PrecisionPoint: PrecisionPoint is now used for move delta to
avoir problem of rounded value caused by zoom level.
* AirResizableEditPolicy:
** This policy already adapts the getMoveCommand for edges with a tree
router. It now also adapts the getMoveCommand for edges with rectlinear
and oblique router to move only the segment "attached" to the move node.
** The getAlignCommand has also been adapted to having the same behavior
as getMoveCommand but on align action.
** The new command SetConnectionBendpointsAccordingToExtremityMove has
been created to handle the edge move.
* SpecificBorderItemSelectionEditPolicy: Same changes as
AirResizableEditPolicy for getMoveCommand
* ArrangeAllWithAutoSize, PinnedElementsLayoutProvider &
AbstractCompositeLayoutProvider: Fix a bug when an arrange selection is
launched. This avoids to have bad size information during the
getMoveCommand and especially in the code added.

Bug: 437097
Change-Id: I2b4b47f2a704b54595d04ea3ce993710389a56d9
Signed-off-by: Laurent Redor <laurent.redor@obeo.fr>