[435363] Correctly use groups in the edit menu
The Copy/PasteLayoutAction are now in the copyLayoutGroup, it was
created before, but actions were added to the copyGroup (for the
contextual menu, ok pour the action bars). 

The delete from diagram was wrongly added in additions before this
copyLayoutGroup, the delete from model was added after the copy layout
group separator. 

Now a deleteFromGroup separator has been added to regroup the two delete
from actions.

Bug: 435363
Change-Id: Idb75fb24936cb10246be5004ef7b23a464a7c064
Signed-off-by: Maxime PORHEL <maxime.porhel@obeo.fr>