[435774] Avoid side-effects on the global package registry
DynamicAcceleoModule.initializeResourceSet() used to call directly
AcceleoPackageRegistry.put(), which delegates to the global EMF package
registry if the package's nsURI is not explicitly registered as dynamic.
When this happens, the global EMF registry is corrupted.

Use AcceleoPackageRegistry.registerEcorePackage() instead, which sets
things up so that we avoid changing the global state.

Bug: 435774
Change-Id: I281cbc9736cc564ca07568a74c128cbee8e75615
Signed-off-by: Pierre-Charles David <pierre-charles.david@obeo.fr>