[435911] Hide obsolete Viewpoint attributes in the properties view
The Viewpoint.{customizes,conflicts,reuses} properties are visible in
the "Advanced" section of the "Viewpoint" type property sheet, but they
were only used by an experimental mode (Movida) which is not maintained
and was never fully supported. Their (undocumented) presence can only
confuse users, so hide them from the UI (until we decide if/when we want
to completely remove them from the code).

Also move the "Icon" property at the bottom of the "General" category
instead of leaving it alone in the "Advanced" category.

Bug: 435911
Change-Id: If11436335222af9f393e33aef85b769a54d2d2e6
Signed-off-by: Pierre-Charles David <pierre-charles.david@obeo.fr>